+++ lib/input/reader/amqp_1.go
@@ -169,6 +169,14 @@ func (a *AMQP1) ReadWithContext(ctx context.Context) (types.Message, AsyncAckFn,
 	if err != nil {
 		if err == amqp.ErrTimeout {
 			err = types.ErrTimeout
+		} else {
+			if dErr, isDetachError := err.(*amqp.DetachError); isDetachError {
+				if dErr.RemoteError != nil {
+					a.log.Errorf("Lost connection due to: %v\n", dErr.RemoteError)
+				}
+				a.disconnect(ctx)
+				err = types.ErrNotConnected
+			}
 		}
 		return nil, nil, err
 	}